// frontend/src/api/generated/models/ClientCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientCreate = {
  name: string;
  industry?: (string | null);
  website_url?: (string | null);
  brand_voice?: (string | null);
  audience?: (string | null);
  eeat_signals?: (string | null);
  language_variant?: string;
  reading_level_target?: (string | null);
  sentence_length_preference?: (string | null);
  banned_words?: Array<string>;
  approved_phrases?: Array<string>;
  oxford_comma?: boolean;
};