// frontend/src/api/generated/models/ClientResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientResponse = {
  name: string;
  industry?: (string | null);
  website_url?: (string | null);
  brand_voice?: (string | null);
  audience?: (string | null);
  eeat_signals?: (string | null);
  language_variant?: string;
  reading_level_target?: (string | null);
  sentence_length_preference?: (string | null);
  banned_words?: Array<string>;
  approved_phrases?: Array<string>;
  oxford_comma?: boolean;
  id: string;
  created_by: string;
  created_at: string;
  updated_at: string;
};